Refactored into a dual-lambda system
Refactored into a dual-lambda system to allow support for deferred responses and follow up messages.

The first lambda (proxy) receives and validate interaction requests from  discord. It then determines what the interaction is and creates an action; a specific instruction intended for the worker lambda. This action is then sent along to the worker lambda.

The second lambda (worker) receives actions from the proxy lambda and then executes the action. This lambda is intended to perform any work that may exceed the 3 second discord response limit. The worker lambda will always respond with a follow up message to the original interaction